fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,12 +5,12 @@
  */
 int main(void)
 {
-	long long int n1 = 0, n2 = 1, n3, i;
+	long int n1 = 0, n2 = 1, n3, i;
 
      	for (i = 2; i <= 50; i++)
      	{
 		n3 = n1 + n2;
-		printf("%lld, ", n3);
+		printf("%ld, ", n3);
 		n1 = n2;
 		n2 = n3;
      	}
